Initial commit for the Locki Trusted Application
This is first version of the Locki TA, where initial development took
place here:
https://github.com/jbech-linaro/optee_examples/tree/attestation

This is still WiP, but has some basic feature working such as creating
users with or without salt, etc. It also contains a couple of other
commands that initially was written for Locki, but those are mostly just
template code not doing a lot. They will probably either be re-worked or
deleted.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>