Initial version of the Locki library
This implements most of the functions supported by the Locki TA, but
just as for the TA, the main focus has been on user creation and
measurements, so the other functions are mostly empty template
functions.

There are also some hardcoded debug flags etc and the shared common
directory and files isn't optimal right now. It works, but should be
better abstracted with proper compile time flags etc.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>