test: add initial test for Locki
This adds some simple testing of the Locki features. Worth noting that
we have an initialize() and an terminate() function in there so support
keeping the TA alive when testing. In real use case this isn't needed,
sicne we plan to set the keep alive attribute on the TA, which means
that once launched it won't terminate until the device reboots. But for
developement and testing purposes, this is needed and convenient.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>