user: add support to leverate TA unique key
By setting the flag USER_TA_UNIQUE_PASSWORD, the Locki TA will also mix
in the TA unique password when deriving the user password upon user
creation. If the user also set the USER_SALT_PASSWORD flag, then either
the user provided salt or the randomly generated salt will also be used
to salt the derviation of the TA unique key. So, with this commit it
possible to get these passwords derived for a user:

 - sha256(provided password | user salt)
 - sha256(provided password | random salt)
 - sha256(provided password | user salt | ta unique key(user salt))
 - sha256(provided password | random salt | ta unique key(user salt))
 - sha256(provided password | ta_unique_key)

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>