locki: implement template code for 'measure'
Note, that running this here doesn't support unauthenticated measure,
likewise it will not really be possible to do measurements here as it is
right now, since the TA is created when invoking this. I.e., there are
no users or anything stored. For this to work, there either needs to be
debug functions (initialize and terminate) available or the TA has to be
configured with the keep alive property. We will address this in later
commits.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>