keys: Implemented the generate_key function
The first draft implementation generated a key and returned it back to
the user. That was never meant how it should be used, it was just some
early test code to check the communication. Now it's been changed to a
instad just simply create a key for you. The key is currently just a 32
byte randomly generated sequence of bytes. In the future, there could be
other ways to generate keys. For example they could be based on the user
password, or the TA unique key or a mix of it etc. Currently key
generation also requires authentication. That could also be changed
later on.

Right now a key is created by first finding a user and a register
element belonging to that user. The id for the key itself is the SHA256
hash of the value of the register and the key handle. This makes the key
id to be derived from various properties that will make it unique to
construct.

Test code has been added for the generate_key call as well.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>