measure: add support for signing measurements
This will sign a register/measurement using an HMAC-SHA256. The key will
be the user's derived password and the data is of course the value of
the register in question.

There are several ways how you can decide to do operations like these.
One could use other generated keys to sign measurements, wrap keys etc.
This servers as a starting point.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>